refactor: Remove scheduleCalendar.ts file
